Inline helpers package into doppler suite
The helpers package was dot-imported in all cases, which makes putting
the functions in a separate package pointless. Also, by moving all test
functions into the suite file, this commit will make future clean-up
easier.
