+++ src/integration_tests/doppler/listener_test.go
@@ -1,8 +1,6 @@
 package doppler_test
 
 import (
-	. "integration_tests/doppler/helpers"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
