+++ src/integration_tests/doppler/syslog_drain_binding_test.go
@@ -5,9 +5,6 @@ import (
 	"net"
 
 	"github.com/nu7hatch/gouuid"
-
-	. "integration_tests/doppler/helpers"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/onsi/gomega/gbytes"
