+++ src/integration_tests/doppler/websocket_streaming_log_test.go
@@ -4,15 +4,12 @@ import (
 	"net"
 	"time"
 
-	. "integration_tests/doppler/helpers"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
 	"github.com/cloudfoundry/dropsonde/factories"
 	"github.com/cloudfoundry/sonde-go/events"
 	"github.com/gorilla/websocket"
 	"github.com/nu7hatch/gouuid"
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
 )
 
 var _ = Describe("Websocket Streaming Logs", func() {
